+++ src/utils.h
@@ -69,7 +69,7 @@ extern double angle_between(double x1, double y1, double z1,
 extern size_t skipspace(const char *s);
 extern void chomp(char *s);
 extern void progress_bar(int val, int total, const char *text);
-extern double poisson_noise(double expected);
+extern int poisson_noise(double expected);
 
 /* Keep these ones inline, to avoid function call overhead */
 static inline struct quaternion invalid_quaternion(void)
